Update README.md
#An Analysis of Kickstarter Campaigns
The project's goal was to analyze Kickstarter data from the past year across the globe across a variety of professions and industries to better deliberate the sweet spot of Kickstarter campaigns. To better aid Louise we narrowed our project's aims by focusing solely on Kickstarters that were in the theater.
![image_name](path/to/OutcomesOfTheater.png)
Here we see that theater has a far larger portion of the Kickstarter. However, as Louise is more specifically a playwright, we should narrow our field to specifically kickstarters that were for plays.
![image_name](path/to/OutcomesOfAllPlays.png)
From these two charts, parsed from our data, we have found that theater-based Kickstarters has fair chances of success, with plays, in particular, having a great amount of success as well. 
It seems that Louise has a good chance of success.